Restructures Locksmith for readability and future modification
As part of code wrangling, the routes have been partioned by use case
as well as supporting  code.
